expirationHandler not being called FIX
Explained:

[NSThread detachNewThreadSelector: spins up a thread and fires the selector. So this is actually enough and what was supposed to be done.

Yet another sync was fired with [self syncingStatusProcessBackgroundTask:nil]; on the next line, causing 2 threads working at the same time on the same task. ExpirationHandler was set twice causing the first one to be overridden so when system called the expirationHandler, the value was nil. Ii's all because the ID and the BGTask are unique, so the system tries to call the first handler and not the 2nd one.

Another possible solution would be to use GCD, like this:
          dispatch_async(dispatch_get_global_queue(QOS_CLASS_BACKGROUND, 0), ^{
            [self syncingStatusProcessBackgroundTask:nil];
          });

This approach leans away from using NSThreads, rely on GCD and spins up exactly one sync with expirationHandler being valid value.

expirationHandler not being called FIX

- cleanup